requirements.txt
streamlit
google-cloud-vision
google-cloud-aiplatform
vertexai
Pillow
requests